requirements.txt
flask
flask-cors
groq 
google-search
python-dotenv
